requirements.txt
django==3.0.3
marshmallow==3.4.0
djangorestframework==3.11.0